backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg[binary]==3.2.12
gunicorn==21.2.0
django-environ==0.11.2
django-cors-headers==4.3.1
whitenoise==6.6.0
drf-spectacular==0.27.2
supabase>=2.9.0
python-dotenv>=1.0.0

